pyproject.toml
[tool.poetry]
name = "myradio"
version = "0.1.0"
description = "MyRadio testing dependencies"
authors = ["Marks Polakovs <marks@markspolakovs.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-playwright = "^0.5.2"
mypy = "^1.11.2"
pytest-mypy = "^0.10.3"
psycopg2 = "^2.9.9"
types-psycopg2 = "^2.9.21.20240819"
psycopg2-binary = "^2.9.9"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
